+++ src/git_workspace/cli/__init__.py
@@ -4,6 +4,7 @@ from git_workspace.cli.callbacks import callback
 from git_workspace.cli.commands import (
     cache_command,
     clone_command,
+    compose_command,
     doctor_command,
     down_command,
     edit_command,
@@ -31,6 +32,7 @@ app = typer.Typer(
 
 app.add_typer(cache_command, name="cache")
 app.add_typer(clone_command)
+app.add_typer(compose_command)
 app.add_typer(doctor_command)
 app.add_typer(down_command)
 app.add_typer(edit_command)